# cmd/docs/v2/swagger.yaml
basePath: /
definitions:
...
  dtos.LoginDTO:
    description: LoginDTO represents a request for user login.
    properties:
      email:
        description: user's email
        example: john.doe@example.com
        type: string
      password:
        description: user's password
        example: strongPassword123
        type: string
    type: object
...
  dtos.RegisterDTO:
    description: RegisterDTO represents a request for user registration.
    properties:
      address:
        description: user's address
        example: 1234 Main St, Springfield
        type: string
      email:
        description: user's email
        example: john.doe@example.com
        type: string
      name:
        description: user's first name
        example: John
        type: string
      password:
        description: user's password
        example: strongPassword123
        type: string
      role:
        description: 'role: lessor, lessee'
        example: lessee
        type: string
    type: object
  dtos.RequestOTPDTO:
    properties:
      email:
        description: user's email
        example: john.doe@example.com
        type: string
    required:
    - email
    type: object
  dtos.ResetPassDTO:
    properties:
      email:
        type: string
      password:
        type: string
      token:
        type: string
    type: object
  dtos.ResetPassRequestDTO:
    properties:
      email:
        example: john.doe@example.com
        type: string
    required:
    - email
    type: object
...
  dtos.VerifyOTPDTO:
    properties:
      email:
        description: user's email
        example: john.doe@example.com
        type: string
      otp:
        description: user's OTP
        example: "123456"
        type: string
    required:
    - email
    - otp
    type: object
  utils.Response:
    properties:
      data: {}
      message:
        example: message
        type: string
      status_code:
        example: 888
        type: integer
    type: object
host: localhost:5000/api/v2
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: API documentation for LeaseEase.
  title: LeaseEase API
  version: "2.0"
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Generates and sends a password reset link to the provided email
        address if the user exists.
      parameters:
      - description: Request payload containing the user's email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.ResetPassRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Reset link sent successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request - Invalid request payload
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found - Email not associated with any account
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error - Failed to send reset email
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Forgot a password
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and set an authentication cookie.
      operationId: "2"
      parameters:
      - description: Login request payload
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dtos.LoginDTO'
      produces:
      - application/json
      responses:
        "201":
          description: User login successfully
          schema:
            items:
              $ref: '#/definitions/utils.Response'
            type: array
        "400":
          description: Invalid request payload
          schema:
            items:
              $ref: '#/definitions/utils.Response'
            type: array
        "500":
          description: Internal server error
          schema:
            items:
              $ref: '#/definitions/utils.Response'
            type: array
      summary: Login an existing user
      tags:
      - auth
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account with the provided details.
      operationId: "1"
      parameters:
      - description: Register request payload
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/dtos.RegisterDTO'
      produces:
      - application/json
      responses:
        "201":
          description: 'User registered successfully"  example({"staus_code": 201,
            "message": "User registered successfully", "data": nil})'
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Invalid request payload
          schema:
            items:
              $ref: '#/definitions/utils.Response'
            type: array
        "500":
          description: Internal server error
          schema:
            items:
              $ref: '#/definitions/utils.Response'
            type: array
      summary: Register a new user
      tags:
      - auth
  /auth/request-otp:
    post:
      consumes:
      - application/json
      description: Sends a one-time password (OTP) to the user's contact information
        provided.
      parameters:
      - description: Request payload containing user identifier
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.RequestOTPDTO'
      produces:
      - application/json
      responses:
        "201":
          description: OTP sent successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data: {}
              type: object
        "400":
          description: Bad Request - Unable to parse request body
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error - Failed to process OTP request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Request OTP for authentication
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Resets the user's password using the provided reset token and new
        password.
      parameters:
      - description: Request payload containing the reset token and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.ResetPassDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request - Invalid request payload
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized - Invalid or expired reset token
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error - Unable to reset password
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Reset user password
      tags:
      - auth
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Validates the OTP provided by the user for authentication.
      parameters:
      - description: Request payload containing OTP and user identifier
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.VerifyOTPDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OTP verification successful
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data: {}
              type: object
        "400":
          description: Bad Request - Invalid OTP payload or incorrect OTP
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error - Failed to verify OTP
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Verify provided OTP
      tags:
      - auth
...
swagger: "2.0"